[Clacc][OpenACC] Fix duplicate diag for acc routine conflict
If an explicit routine directive on a function definition is discarded
by `ActOnOpenACCRoutineDirective` because it conflicted with a
previous routine directive, the `ACCRoutineDeclAttr` left on the
function definition still has the `Inherited` property because it
logically isn't replaced by the new routine directive.  However,
without this patch, `ActOnOpenACCRoutineDirective` assumes that
property means it hasn't yet processed the new routine directive, so
it executes again and duplicates the conflict diagnostic.  Instead of
using the `Inherited` property, this patch uses the directive stack to
record whether an explicit routine directive has been acted on.