[Clacc][OpenACC][C++] Fix ctor transform in OpenACC construct
Without this patch, when `TransformACCToOMP` runs, it sometimes
produces spurious complaints that default constructors don't exist at
calls to them within the transformed OpenACC construct.  To fix that,
this patch modifies the variable initialization transformation in
`TransformContext::TransformDefinition` to mimic
`Sema::InstantiateVariableInitializer`: it adjusts the initialization
style argument to `TransformInitializer` so that it doesn't complain
that there is no default constructor, and then it calls
`ActOnUninitializedDecl` instead of `AddInitializerToDecl`, which
would otherwise seg fault for this case.  It also adjusts the last
argument to `AddInitializerToDecl` to match the call in
`Sema::InstantiateVariableInitializer`, but we don't yet have a case
that fixes.