[Clacc][lit] Implement DEFINE and REDEFINE directives
These directives define per-test lit substitutions.  The concept was
discussed at
<https://discourse.llvm.org/t/iterating-lit-run-lines/62596/10>.

For example, in a test file, the following directives define a
`%{check}` substitution as well as empty default values for
substitutions that serve as parameters of `%{check}`:

```
// DEFINE: %{cflags} =
// DEFINE: %{fc-prefix} =

// DEFINE: %{check} = %clang_cc1 %{cflags} -emit-llvm -o - %s | \
// DEFINE:            FileCheck -check-prefix=%{fc-prefix} %s
```

The following directives then redefine the parameters before each use
of `%{check}`:

```
// REDEFINE: %{cflags} = -foo
// REDEFINE: %{fc-prefix} = FOO
// RUN: %{check}

// REDEFINE: %{cflags} = -bar
// REDEFINE: %{fc-prefix} = BAR
// RUN: %{check}
```

Of course, `%{check}` would typically be more elaborate, increasing
the benefit of the reuse.

One issue is that the strings `DEFINE:` and `REDEFINE:` already appear
in 5 tests.  This patch adjusts those tests.  Alternatively, maybe lit
keywords from now on should start with `LIT-` to avoid collisions.  In
this case, we might use `LIT-DEF:` and `LIT-REDEF:`.

This patch also expands the documentation on existing lit substitution
behavior.

We are attempting to upstream this patch in D132513.