[Clacc][lit] Hack together %for/%data/(RE)DEFINE integration
That is, make it possible to loop over DEFINE/REDEFINE using %for.
That integration helped while migrating Clacc's tests suites from
using %for/%data to using DEFINE/REDEFINE to implement loops of RUN
lines.  The full migration will appear later in a single commit, so no
commit ever captures a state where the integration was used.
Moreover, a commit after that will discard %for/%data altogether.
Nevertheless, we don't want to purge the integration from Clacc's
history in case it ever proves useful to revisit.  In that case, keep
in mind that this integration was never well tested.  It was just
enough to get us through the migration.