[Clacc][OpenACC] Translate routine directives in functions
This patch completes initial support for `routine` directives within
functions by fixing their translation to OpenMP, documenting that
support, and finishing testing it.  Most of the patch focuses on
handling the case that an explicit `routine` directive is discarded
during translation.  The patch also tweaks the implementation of the
similar case of `ACCDirectiveStmt` for consistency with this one.