[Clacc][OpenACC] Fix inherited no_create for ptr-based subarray
That is, when `no_create(ptr[S:L])` on a `data` construct suppresses
the implicit data attribute for `ptr` on an enclosed `parallel`
construct, fix the `map` clause (with "inherited" `ompx_no_alloc`)
added in the translation of the latter so that it specifies `ptr[S:L]`
not just `ptr`.

In the case that the subarray is already present, `ptr` is treated as
unmapped in the `parallel` construct without this fix, resulting in
memory access errors where the subarray is accessed.  With this fix,
`ptr` is instead treated as the base of the already mapped subarray.

In the case that the subarray is not already present, we assume that
accesses to the subarray within the `parallel` construct are
unreachable, so this fix is irrelevant.  This patch adds a note about
this assumption to the status doc as the assumption appears not to be
consistent with the OpenACC spec.

Also, stop generating `defaultmap(tofrom:scalar)` for pointers as it's
irrelevant to them.