[Clacc][OpenACC] Fix suppressing DMA for pointer variable
For example:

```
 #pragma acc data copy(ptr)
 #pragma acc parallel
 ptr += 1;
```

Without this fix, the above fails to copy the new value of `ptr` back
to the host because it relies on OpenMP's implicit `map` clause for
`ptr` on the `parallel` construct, and that implicit `map` clause
treats `ptr` is the base pointer of a subarray.  With this fix, Clacc
adds an explicit `map` clause to override that behavior.

This issue only affects the case where OpenACC's implicit data
attribute for `ptr` on the `parallel` construct is suppressed by a
non-`no_create` explicit data clause on the `data` construct.