[Clacc][CI][Kokkos] Use a stable Kokkos fork
For now, we need a Kokkos fork that has a development version of the
OpenACC backend.  However, the one we previously chose sometimes
experiences history rewrites, so it isn't stable enough for Clacc's CI
testing.  For now, we've created our own fork.  Another advantage is
that the new fork contains our Clacc-specific changes so we don't have
to maintain a kokkos.patch here anymore.