[Clacc][OpenACC][C++] Improve lambda analysis
This patch fixes many cases where the OpenACC analysis incorrectly
looked beyond an enclosing lambda to find the enclosing function, data
clauses, loop directives, routine directives, etc.  As a result, this
patch fixes various assertion failures, incorrect or missed
diagnostics, and other misbehaviors involving lambdas.  Generally, it
causes lambda definitions nested within other function definitions to
behave more like function definitions that are not nested.

This patch also eliminates a spurious type conversion diagnostic when
assigning a lambda to an auto variable within an OpenACC construct.
The problem was that TransformContext did not properly transform the
type of the variable during the translation to OpenMP.

There is more work to be done for lambdas.  That is, there might be
some lambda issues this patch does not fix, and it is not clear in
every case if the above changes always follow the intention of the
OpenACC specification, which might require some clarification.