[Clacc][OpenACC] Remove source-to-source example in README.md
In the "Basic Usage" section of the top-level `README.md`, the example
output for source-to-source mode has inadvertently long been out of
date with respect to the preceding input test code.  Moreover, in
source-to-source mode, that input test code triggers an
error-by-default warning about its use of an OpenMP extension
(`ompx_hold`).  It seems challenging to find a simple example that
won't trigger that warning and that will behave correctly on both
NVIDIA GPUs and AMD GPUs (given that the latter doesn't have printf
support yet).

Generally, because source-to-source mode has many caveats like this,
it seems best not to discuss it under "Basic Usage".  Traditional
compilation is Clacc's main goal, so it's sufficient here.