[Clacc][OpenACC] Fix ompStmtPrintsDifferently if indent differs
The problem is indentation can be different when printing an
associated statement from an OpenACC node versus its OpenMP node, but
that's not a reason to force the OpenACC and OpenMP versions to be
printed separately.  fd0d534a934d previously fixed this problem for
immediate child combined directives, but it didn't handle combined
directives deeper in the associated statement.  This patch should
handle all cases by just turning off indentation altogether when
comparing OpenACC and OpenMP versions.