[Clacc][OpenACC] Simplify AST print indentation
Without this patch, `-fopenacc-print` and `-fopenacc-ast-print` try to
align indentation of consecutive OpenMP directives and sometimes their
associated statement that are translated from OpenACC constructs.
While that indentation is typical in input source code, it sometimes
creates inconsistencies with Clang's normal AST printing style, in
which consecutive directives and their associated statements are
progressively indented.  For example, it seems
`-fopenacc-ast-print=omp` should always follow normal AST printing
style.  More importantly, this behavior complicates the
implementation.  In particular, a future commit will implement the
`tile` clause, which will add the new case of a single OpenACC
effective directive translated to multiple OpenMP directives, and that
will further complicate the implementation's tracking of OpenMP
directives to be aligned.

This patch eliminates alignment in many cases.  After this patch, the
strategy of `-fopenacc-ast-print` is to align only where necessary: it
aligns consecutive OpenMP directives only where printing both the
OpenACC and OpenMP directives (one within comments) followed by a
single copy of the associated statement, which is further indented.
If the associated statement must be printed separately for both, or if
only OpenMP is printed, then there is no need to align the OpenMP
directives.  The strategy of `-fopenacc-print` is to align except
where AST printing is required: it tries to align consecutive OpenMP
directives and their associated statement except when it cannot reuse
the original OpenACC directive's associated statement from the source
code.