[Clacc][OpenMP] Fix libomptarget test mapping/ompx_hold/struct.c
For me, the test fails for nvptx64 offload.  The problem was
introduced by D146838, which landed as 747af2415519.  It tries to copy
a string constant's address from device to host and then print the
string.  This patch copies the contents of the string instead.

We upstreamed this previously as D149623, which landed as
3a1e06e0e82c.

This patch also removes the test from `LIT_FILTER_OUT` throughout
Clacc's CI config.