[Clacc][CI][Kokkos] Update Kokkos
This patch adjusts CI to use a version of Kokkos that is nearly
identical to a 2023/08/17 version of upstream Kokkos's develop branch,
which includes our changes to support compiling the OpenACC backend
with Clacc, so we don't have to maintain those changes locally
anymore.

The new Kokkos version to be used by CI also includes changes that
have not yet been upstreamed.  First, like the previous version used,
it includes changes in the OpenACC backend to support targeting the
host, a scenario that we exercise on CI hosts without GPUs.  Second,
it includes new experimental changes not previously tested in this CI
to improve performance when compiling with Clacc.  Finally, it
excludes some additional tests that fail.