[Flang][WIP/RFC] Enable TSan for Flang
This patch enables ThreadSanitizer analysis for Fortran codes compiled with Flang.
The patch is marked as WIP/RFC since it is at the moment a proof of concept.

This patch as is cannot be the final solution, as it disables the blacklisting of
functions and any no-sanitize function attribute applied in C/C++

Under review as #74643