[Clacc][CI] Drop ulimit -t
Some time ago, ExCL admins fixed the gitlab runner installation to
kill processes left behind by a CI job when the job terminates.  That
should eliminate the need for `ulimit -t`, which was problematic in a
few ways.  First, it is challenging to use correctly because it
doesn't track real time but instead CPU usage, which can be none in
the case of no host activity or multiplied in the case of CPU
multithreading.  Second, it doesn't produce an obvious timeout
diagnostic.  In the case of lit tests, we'll now rely more on lit's
`--timeout` feature, which avoids both of those problems.