[Clacc] Work around github README.md bug
If files with names like `*LICENSE*` or `*SECURITY*` (I'm guessing at
the exact pattern) are present, then where github normally renders
`README.md`, it also shows tabs for these additional files.  The
problem is that, when using tabs, it malfunctions and uses the
versions of those files from the repo's default branch instead of the
currently selected branch.

That's problematic for Clacc where `README.md` provides Clacc-specific
information.  Work around it by renaming the problematic files so
github doesn't see them.

This problem did not always occur, and the tabs functionality appears
to be relatively new:

- https://github.com/orgs/community/discussions/86658
- https://github.com/orgs/community/discussions/82712